pyproject.toml
[project]
name = "youtube-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.13.1",
    "youtube-transcript-api>=1.2.2",
    "tavily-python>=0.3.0",
]
